# server/openapi.yaml
openapi: 3.0.0
info:
  title: Neko API Reference
  version: "3.0.0"
  description: |
    Neko uses WebRTC and WebSocket besides REST API to provide real-time video streaming, interactive collaboration, and remote desktop functionality.

    ### Core Technologies

    - **Video Streaming over WebRTC**
      - Neko leverages **WebRTC** for real-time, low-latency, high-quality video and audio streaming.
      - Keyboard and mouse events are transmitted over the same WebRTC connection for seamless interaction.

    - **Real-Time Signaling and Communication over WebSocket**
      - **WebSocket** is used for sSignaling in WebRTC (e.g., session establishment, ICE candidate exchange).  
      - Live chat, session updates, and keyboard/mouse fallback are handled over WebSocket.

    - **REST API for Everything Else**
      - **Authentication & Session Management** - Secure user authentication and session handling.  
      - **Room and User Management** - Creating, modifying, and controlling user access to rooms.  
      - **Desktop Controls** - Adjusting screen configurations, managing keyboard and mouse events.
      - **System Health & Metrics** - Retrieving system health checks and operational insights.  

    ### How It All Works Together

    1. **WebRTC** handles media streaming for **real-time video, audio and data**.
    2. **WebSocket** ensures a **persistent, bidirectional** connection for low-latency interactions.  
...

...
  license:
    name: Apache 2.0
...
servers:
...
paths:
...
  /api/profile:
    post:
      tags:
        - current-session
      summary: Update Profile
      description: Update the current user's profile without syncing it with the member profile (experimental).
      operationId: profile
      responses:
        '204':
          description: Profile updated successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberProfile'
        required: true

  #
  # sessions
  #
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: NEKO_SESSION
      description: Authentication using a session cookie.
    BearerAuth:
      type: http
      scheme: bearer
      description: Authentication using a Bearer token.
    TokenAuth:
      type: apiKey
      in: query
      name: token
      description: Authentication using a token passed as a query parameter.
  responses:
...
    Unauthorized:
      description: The request requires user authentication.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
...
  schemas:
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
          description: Detailed error message.
...
    MemberProfile:
      x-tags:
        - members
      type: object
      properties:
        name:
          type: string
          description: The name of the member.
        is_view_only:
          type: boolean
          description: Indicates that the member is passive and all interactive capabilities are denied server-side.
        is_admin:
          type: boolean
          description: Indicates if the member is an admin.
        can_login:
          type: boolean
          description: Indicates if the member can log in.
        can_connect:
          type: boolean
          description: Indicates if the member can connect.
        can_watch:
          type: boolean
          description: Indicates if the member can watch.
        can_host:
          type: boolean
          description: Indicates if the member can host.
        can_share_media:
          type: boolean
          description: Indicates if the member can share media.
        can_access_clipboard:
          type: boolean
          description: Indicates if the member can access the clipboard.
        sends_inactive_cursor:
          type: boolean
          description: Indicates if the member sends inactive cursor.
        can_see_inactive_cursors:
          type: boolean
          description: Indicates if the member can see inactive cursors.
        plugins:
          type: object
          additionalProperties: true
          description: Additional plugin settings.
...
security:
  - BearerAuth: []
  - CookieAuth: []
  - TokenAuth: []